+++ tests/test_visitors/conftest.py
@@ -14,10 +14,23 @@ from wemake_python_styleguide.visitors.base import BaseVisitor
 @pytest.fixture(scope='session')
 def assert_errors():
     """Helper function to assert visitor violations."""
-    def factory(visitor: BaseVisitor, errors: Sequence[str]):
-        assert len(errors) == len(visitor.violations)
+    def factory(
+        visitor: BaseVisitor,
+        errors: Sequence[str],
+        ignored_types=None,
+    ):
+        if ignored_types:
+            real_errors = [
+                error
+                for error in visitor.violations
+                if not isinstance(error, ignored_types)
+            ]
+        else:
+            real_errors = visitor.violations
 
-        for index, error in enumerate(visitor.violations):
+        assert len(errors) == len(real_errors)
+
+        for index, error in enumerate(real_errors):
             assert error.code == errors[index].code
             if isinstance(error, (ASTViolation, TokenizeViolation)):
                 assert error._node is not None  # noqa: Z441